GP.5.1 audit: fix lint warnings + stale codemaps, harden test coverage
Independent audit of the GP.5.1 fee-split changes.  No correctness or
security defect found in the contract logic: conservation
(userAmount + poolAmount = msg.value, exact), the budget clamp, the
unchecked-subtraction soundness (poolAmount <= floor(v/2) since
maxFeeBps <= 5000), replay resistance (deploymentId + per-depositor
nonce + eight-field hash), the zero-deposit guard, and the constructor
guards were all re-verified from first principles.  Two material
defects were found and fixed, plus polish:

- Lint: the PR introduced 7 `unsafe-typecast` warnings (the repo had
  zero un-suppressed).  Fixed the one production cast at
  KnomosisBridge.depositETHWithFee with a documented
  forge-lint suppression (the value is provably <= 10^12 < 2^63 in
  that branch); widened FeeSplitMath's `split` / `receiptHash` input
  params to uint256 so the 5 cross-check call-site casts disappear
  (abi.encode of an in-range uint256 is byte-identical to the
  contract's uint64 encoding, so the receiptHash stays byte-equal);
  and suppressed the one remaining uint64-budgetGrant cast in
  FeeSplitMath.  Whole-project `forge lint` is now clean.

- Codemaps: the session-1 commit shipped stale codemaps.
  regenerate_codemaps.py enumerates sources via `git ls-files`, and the
  regen had run before `git add`, so the four new files
  (DepositFeeSplit.lean, BridgeFeeSplit.t.sol, DepositFeeSplit.t.sol,
  FeeSplitMath.sol) were untracked and silently skipped -- the
  committed codemaps were missing ~820 lines of their declarations,
  which would have failed the CI codemap gate.  Regenerated with the
  files tracked; idempotent.

- Coverage: +3 behavioural tests (37 -> 40) -- the migration
  circuit-breaker on depositETHWithFee, the minFeeBps==maxFeeBps==0
  forced-zero-fee deployment, and the shared per-depositor nonce across
  depositETH / depositETHWithFee.

- Docs: corrected a stale FeeSplitMath docstring (claimed the clamp
  constant is read from the contract; it is a mirrored literal), fixed
  the Lean randomised-bound notation to the inclusive ranges the
  generators produce ([1, 2^96], [1, 2^50]), and synced the 37 -> 40
  count across CLAUDE.md / AGENTS.md / the plan.

lake test green (fixture byte-stable); clean-rebuild 458-test forge
suite green; whole-project forge lint clean; six Lean audit gates pass.

+++ solidity/test/utils/FeeSplitMath.sol
@@ -15,9 +15,11 @@ pragma solidity ^0.8.20;
 ///         against this reference via `vm.expectEmit`.  Together they
 ///         establish `contract == reference == Lean spec` without any
 ///         single file checking the formula against itself.  The clamp
-///         constant is read from `KnomosisBridge.MAX_BUDGET_PER_DEPOSIT`
-///         so the reference can never silently drift from the
-///         deployed cap.
+///         constant is a mirrored literal (a contract's public constant
+///         is not reachable via the type name from another file); both
+///         test suites assert it equals the deployed
+///         `KnomosisBridge.MAX_BUDGET_PER_DEPOSIT` getter and the Lean
+///         fixture header, so any drift fails loudly.
 library FeeSplitMath {
     /// @notice Per-deposit budget-grant ceiling.  MUST equal
     ///         `KnomosisBridge.MAX_BUDGET_PER_DEPOSIT`; the test suites
@@ -39,35 +41,49 @@ library FeeSplitMath {
     ///           budgetGrant = min(floor(poolAmount / weiPerBudgetUnit),
     ///                             MAX_BUDGET_PER_DEPOSIT)
     ///
-    /// @dev    Caller must ensure `feeBps <= 10000` (so
-    ///         `poolAmount <= v` and the subtraction does not underflow)
-    ///         and `weiPerBudgetUnit >= 1` (so the division is defined).
+    /// @dev    `feeBps` and `weiPerBudgetUnit` are taken as `uint256`
+    ///         (rather than the contract's `uint16` / `uint64`) so call
+    ///         sites that parse JSON values can pass them without a
+    ///         narrowing cast — the arithmetic is value-identical.
+    ///         Caller must ensure `feeBps <= 10000` (so `poolAmount <=
+    ///         v` and the subtraction does not underflow) and
+    ///         `weiPerBudgetUnit >= 1` (so the division is defined).
     ///         Both hold for every admissible on-chain call: the
     ///         contract rejects `feeBps > maxFeeBps <= MAX_FEE_BPS_CAP
     ///         = 5000` and rejects `weiPerBudgetUnit = 0` at
-    ///         construction.
-    function split(uint256 v, uint16 feeBps, uint64 weiPerBudgetUnit)
+    ///         construction.  `budgetGrant` is returned as `uint64`
+    ///         (the contract's type and the event field width).
+    function split(uint256 v, uint256 feeBps, uint256 weiPerBudgetUnit)
         internal
         pure
         returns (uint256 userAmount, uint256 poolAmount, uint64 budgetGrant)
     {
-        poolAmount = (v * uint256(feeBps)) / 10_000;
+        poolAmount = (v * feeBps) / 10_000;
         userAmount = v - poolAmount;
-        uint256 raw = poolAmount / uint256(weiPerBudgetUnit);
+        uint256 raw = poolAmount / weiPerBudgetUnit;
+        // casting to `uint64` is safe: the ternary takes this branch
+        // only when raw <= MAX_BUDGET_PER_DEPOSIT = 10^12 < 2^63.
+        // forge-lint: disable-next-line(unsafe-typecast)
         budgetGrant = raw > uint256(MAX_BUDGET_PER_DEPOSIT) ? MAX_BUDGET_PER_DEPOSIT : uint64(raw);
     }
 
     /// @notice Recompute the canonical fee-split `receiptHash`.
     ///         Mirrors `KnomosisBridge._registerDepositWithFee`.
+    /// @dev    `resourceId`, `budgetGrant`, and `nonce` are taken as
+    ///         `uint256` (the contract uses `uint64`).  ABI-encoding an
+    ///         integer `< 2^64` as `uint256` yields the identical
+    ///         32-byte word the contract's `uint64` `abi.encode`
+    ///         produces, so the hash is byte-equal; widening lets call
+    ///         sites avoid narrowing casts.
     function receiptHash(
         bytes32 deploymentId,
         address sender,
-        uint64 resourceId,
+        uint256 resourceId,
         address token,
         uint256 userAmount,
         uint256 poolAmount,
-        uint64 budgetGrant,
-        uint64 nonce
+        uint256 budgetGrant,
+        uint256 nonce
     ) internal pure returns (bytes32) {
         return keccak256(
             abi.encode(
